requirements.txt
# Core dependencies
click>=8.1.0
ffmpeg-python>=0.2.0
pydantic>=2.0.0
pyyaml>=6.0
tqdm>=4.65.0

# Web UI dependencies (optional)
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6

# Development dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0

# Optional dependencies for advanced features
opencv-python>=4.8.0  # For scene detection
numpy>=1.24.0  # For audio analysis